chore: swap out old page for enhanced
