--- web-ui/src/pages/CheckinsReportEnhancedPage.css
@@ -1,27 +1,0 @@
-.checkins-report-page {
-  display: flex;
-  flex-direction: column;
-  margin: 1rem;
-
-  .checkins-report-page-reports {
-    display: flex;
-    flex-direction: column;
-    margin-top: 1rem;
-  }
-
-  .checkins-report-page-dates {
-    display: flex;
-    justify-content: flex-end;
-    text-align: right;
-    gap: 2rem;
-    margin: 0;
-  }
-
-  .checkins-report-page-no-data {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    height: 10rem;
-  }
-}
